refactor(data): remove finetuning compatibility module
