--- src/megatron/bridge/data/finetuning.py
@@ -1,20 +1,0 @@
-# Copyright (c) 2025, NVIDIA CORPORATION.  All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""Compatibility imports for finetuning data handling utilities."""
-
-from megatron.bridge.data.batch_utils import prepare_finetuning_batch, split_batch_into_microbatches
-
-
-__all__ = ["prepare_finetuning_batch", "split_batch_into_microbatches"]
