+++ tests/functional_tests/test_groups/data/test_samplers.py
@@ -446,7 +446,7 @@ class TestFinetuningUtilities:
         """Test basic batch splitting functionality."""
         import torch
 
-        from megatron.bridge.data.finetuning import split_batch_into_microbatches
+        from megatron.bridge.data.batch_utils import split_batch_into_microbatches
 
         batch = {
             "tokens": torch.randint(0, 1000, (16, 240)),
@@ -466,7 +466,7 @@ class TestFinetuningUtilities:
         """Test that splitting preserves data integrity."""
         import torch
 
-        from megatron.bridge.data.finetuning import split_batch_into_microbatches
+        from megatron.bridge.data.batch_utils import split_batch_into_microbatches
 
         batch = {
             "tokens": torch.arange(16 * 10).reshape(16, 10),
@@ -486,7 +486,7 @@ class TestFinetuningUtilities:
         """Test splitting with mixed tensor and non-tensor items."""
         import torch
 
-        from megatron.bridge.data.finetuning import split_batch_into_microbatches
+        from megatron.bridge.data.batch_utils import split_batch_into_microbatches
 
         batch = {
             "tokens": torch.randint(0, 1000, (16, 240)),
@@ -507,7 +507,7 @@ class TestFinetuningUtilities:
         """Test prepare_finetuning_batch function."""
         import torch
 
-        from megatron.bridge.data.finetuning import prepare_finetuning_batch
+        from megatron.bridge.data.batch_utils import prepare_finetuning_batch
 
         # Create mock dataloader that yields global batches
         def mock_dataloader():
@@ -544,7 +544,7 @@ class TestFinetuningUtilities:
         """Test with different seq_lengths across global batches."""
         import torch
 
-        from megatron.bridge.data.finetuning import prepare_finetuning_batch
+        from megatron.bridge.data.batch_utils import prepare_finetuning_batch
 
         # Mock dataloader with variable seq_lengths
         def mock_dataloader():
